app/requirements.txt
Flask
Jinja2
python-dotenv
requests
gunicorn
prometheus-flask-exporter
